Add Horizontal Dividers
Added horizontal dividers to the player and creator. These are intended to be used mainly for fractions, but can also be used for visual purposes. This commit also includes a change to Slides, which now just stores steps instead of a full instructions object for each step. This fixes a bug where content could only be used in the step that was active when the content was created.