Finish implementation of exponents and subscripts
All kinds of container now behave themselves when inside an exponent. This allows fractional exponents, and nested exponents. All containers function properly in the creator as well when inside an exponent. Previously, clicking on the inside border of an HBox would not work. Additionally, layout/sizing of terms has been tweaked. The creator gets a new option for the SubSuper container, where the alignment of the exponent can be changed. This allows for large exponents to look right.