Change canvas resize behavior
Canvases no longer animate between sizes, as this affected performance. Instead they instantly become the required size. If a layout is too wide, the canvas will now automatically resize to accomodate this and scroll horizontally. This will allow for a larger font size on the smallest screens while accomodating the same layouts that would fit on a larger screen.