Update the loader
The loader is now exported as either an AMD module or a global variable. It can now load every container with a certain class name as before, or load individual ones in varying ways. Particularly, the data-src attribute can be used on a container to specify where its file should be loaded from. Also, support fixing the height of the container on IE10.