HBoxes vertically align things more accurately
Before this change, SubSupers/RootContainers would exaggerate their height so that their main text was vertically centered. This looked off if, for example, a SubSuper had a very tall upper part. In this case, it would have a large amount of empty space on the bottom. This is no longer the case, so layouts like this look better.